+++ src/handler/events/readyHandler.ts
@@ -3,15 +3,13 @@ import type Wrapper from '../structures/wrapper';
 import { concatMap, fromEvent, type Observable, take } from 'rxjs';
 import * as Files from '../utilities/readFile';
 import { errTap, scanModule } from './observableHandling';
-import { CommandType } from '../structures/enums';
-import { SernError } from '../structures/errors';
+import { CommandType, SernError, type ModuleStore } from '../structures';
 import { match } from 'ts-pattern';
 import { Result } from 'ts-results-es';
 import { ApplicationCommandType, ComponentType } from 'discord.js';
 import type { CommandModule } from '../../types/module';
 import type { Processed } from '../../types/handler';
 import type { ModuleManager } from '../contracts';
-import type { ModuleStore } from '../structures/moduleStore';
 import { _const, err, ok } from '../utilities/functions';
 import { defineAllFields } from './operators';
 import SernEmitter from '../sernEmitter';
