+++ src/handler/structures/index.ts
@@ -1,6 +1,6 @@
 import Context from './context';
 import type Wrapper from './wrapper';
 import { ModuleStore } from './moduleStore';
-
+export * from './errors';
 export * from './enums';
 export { Context, Wrapper, ModuleStore };
